requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.8
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.15.1
idna==3.7
PyJWT==2.8.0
PyMySQL==1.1.0
pypng==0.20220715.0
qrcode==7.4.2
redis==5.0.4
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
